Last post we did for text classification with CNN (#5227)
Last post with text classification with CNNs at character level. The code is available [here](https://github.com/Azure/Cortana-Intelligence-Gallery-Content/tree/master/Tutorials/Deep-Learning-for-Text-Classification-in-Azure). It's in python and R. If you guys want me to push the code as one of the examples let me know.

We have a web app that shows the system, you can play with it!:
http://osdcwebappdeeplearning.azurewebsites.net/