Fix data layout handling in BucketSentenceIter (Fixes #5509) (#6707)
* Fix data layout handling in BucketSentenceIter (Fixes #5509)

BucketSentenceIter did not correctly specify the used data layout in the
provide_data and provide_label attributes. This lead to the executor_group
always assuming an NCHW layout, trying to split up the time dimension instead of
the batch_size dimension over all contexts in case that TNC layout was used.

For some reason this nevertheless worked for the provided FusedRNNCell,
resulting in model parallelism.

Interestingly this patch reduces performance by 25% as it replaces the unmeant
model parallelism with the "correct" data parallelism. (Reference architecture
is a simple single-layer CharRNN trained on 4 GTX 1080 cards. Without the patch
~2000 samples/sec, with the patch ~1500 samples/sec)

* BucketSentenceIter: Correct default layout

* BucketSentenceIter: Fix lint