Pip installer for CoreML Converter: mxnet-to-coreml (#7624)
* Fixing CoreML converter's README: typos/grammar/etc.

* CoreML converter README update: Talk about layers first and then about models.

* Providing examples on converting various standard models; calling out issues with InceptionV3.

* Fixing CoreML converter's README: typos/grammar/etc.

* CoreML converter README update: Talk about layers first and then about models.

* Providing examples on converting various standard models; calling out issues with InceptionV3.

* Pip installer for converter: mxnet-coreml-converter.

Runs only on MacOS and python 2.7. Once inside the directory pip_package, user needs
to run:
python setup.py bdist_wheel
twine upload dist/*

Once uploaded it'll look like this:
https://testpypi.python.org/pypi/mxnet-coreml-converter

Also updated the README for converter to reflect this.

Note that we are going with a package per tool for the time being. Please leave feedback if you think it is better to adopt the policy of all the tools in one single package.

Unit tests continue to pass.

* More informative pypi package documentation.

* Updating MacOS in release notes to 10.11 after testing on it.

* Changing the name to mxnet-to-coreml and version to 0.1.0.

* Added license to setup.py

* Updating readme files with the correct pip package name.