Implement Khatri-Rao operator (#7781)
* Implement Khatri-Rao operator

Create an operator for the Khatri-Rao function implemented in #6567. The
operator accepts a variable number of input matrix arguments as well as a flag
indicating whether the Khatri-Rao product should be computed row-wise or
column-wise. This option is provided because the row-wise operation is more
memory efficient.

* Implement Khatri-Rao operator

Remove row-wise flag from parameter/keyword set. That is, only compute the
column-wise Khatri-Rao product.

* Implement Khatri-Rao operator

Remove unnecessary type-specific `flip` code. The fully-templatized version
handles the `float` and `double` instances.